Seperated cli component in scripts
